requirements.txt
requests>=2.31.0
pandas>=2.0.0
mplfinance>=0.12.0
matplotlib>=3.7.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
python-dotenv>=1.0.0
ccxt>=4.0.0

